Update to ChangeSet 196

+++ Richman4L/ExternalLibrary/FIGlet.Net/FIGlet.Net/FigletFont.cs
@@ -150,34 +150,34 @@ namespace WenceyWang .FIGlet
 				throw new ArgumentOutOfRangeException ( nameof ( line ) ) ;
 			}
 
-			if ( Lines [ Convert . ToByte ( sourceChar ) ] == null )
+			if ( Lines [ Convert . ToInt16 ( sourceChar ) ] == null )
 			{
 				return string . Empty ;
 			}
 
-			return Lines [ Convert . ToByte ( sourceChar ) ] [ line ] ;
+			return Lines [ Convert . ToInt16 ( sourceChar ) ] [ line ] ;
 		}
 
-		//private void LoadLines ( List<string> fontLines )
-		//{
-		//	Lines = fontLines;
-		//	string configString = Lines . First ( );
-		//	string [ ] configArray = configString . Split ( ' ' );
-		//	Signature = configArray . First ( ) . Remove ( configArray . First ( ) . Length - 1 );
-		//	if ( Signature == "flf2a" )
-		//	{
-		//HardBlank = configArray . First ( ) . Last ( ) . ToString ( );
-		//Height = configArray . GetIntValue ( 1 );
-		//BaseLine = configArray . GetIntValue ( 2 );
-		//MaxLength = configArray . GetIntValue ( 3 );
-		//OldLayout = configArray . GetIntValue ( 4 );
-		//CommentLines = configArray . GetIntValue ( 5 );
-		//PrintDirection = configArray . GetIntValue ( 6 );
-		//FullLayout = configArray . GetIntValue ( 7 );
+		//	}
 
 		//CodeTagCount = configArray . GetIntValue ( 8 );
+		//FullLayout = configArray . GetIntValue ( 7 );
+		//PrintDirection = configArray . GetIntValue ( 6 );
+		//CommentLines = configArray . GetIntValue ( 5 );
+		//OldLayout = configArray . GetIntValue ( 4 );
+		//MaxLength = configArray . GetIntValue ( 3 );
+		//BaseLine = configArray . GetIntValue ( 2 );
+		//Height = configArray . GetIntValue ( 1 );
+		//HardBlank = configArray . First ( ) . Last ( ) . ToString ( );
+		//	{
+		//	if ( Signature == "flf2a" )
+		//	Signature = configArray . First ( ) . Remove ( configArray . First ( ) . Length - 1 );
+		//	string [ ] configArray = configString . Split ( ' ' );
+		//	string configString = Lines . First ( );
+		//	Lines = fontLines;
+		//{
 
-		//	}
+		//private void LoadLines ( List<string> fontLines )
 		//}
